rename server etc.
