+++ scripts/move-binary.js
@@ -1,23 +1,23 @@
-const execa = require('execa')
-const fs = require('fs')
+const execa = require('execa');
+const fs = require('fs');
 
-let extension = ''
+let extension = '';
 if (process.platform === 'win32') {
-  extension = '.exe'
+    extension = '.exe';
 }
 
 async function main() {
-  const rustInfo = (await execa('rustc', ['-vV'])).stdout
-  const targetTriple = /host: (\S+)/g.exec(rustInfo)[1]
-  if (!targetTriple) {
-    console.error('Failed to determine platform target triple')
-  }
-  fs.renameSync(
-    `server${extension}`,
-    `server-${targetTriple}${extension}`
-  )
+    const rustInfo = (await execa('rustc', ['-vV'])).stdout;
+    const targetTriple = /host: (\S+)/g.exec(rustInfo)[1];
+    if (!targetTriple) {
+        console.error('Failed to determine platform target triple');
+    }
+    fs.renameSync(
+        `rbk_server${extension}`,
+        `rbk_server-${targetTriple}${extension}`
+    );
 }
 
 main().catch((e) => {
-  throw e
-})
+    throw e;
+});
